requirements.txt
google-generativeai
langchain-google-genai
langgraph
langchain-core
streamlit
python-dotenv